refactor: split AgentStatus into focused submodules and tests
