--- console/src/AgentStatus.tsx
@@ -1,13 +1,0 @@
-export {
-  AgentStatus,
-  createAgentStatusRows,
-  createAgentStatusSections,
-  parseAgentStateResponse,
-  createMockAgentStateResponse,
-  isAgentStateMockQueryEnabled,
-  shouldUseMockAgentState,
-  startAgentStateAutoRefresh,
-  AGENT_STATE_MOCK_NOTICE_MESSAGE,
-  AGENT_STATE_REFRESH_INTERVAL_MS,
-  shouldShowAgentStatusErrorForEventSourceError,
-} from "./AgentStatus/index";
