+++ console/src/AgentStatus/GameStatusSection.tsx
@@ -1,10 +1,10 @@
 import { AgentStatusSectionCard } from "./AgentStatusSectionCard";
-import type { AgentStatusRow } from "./AgentStatusSectionCard";
+import type { AgentStatusRow } from "./types";
 
 export const GAME_SECTION_TITLE = "ゲームの状態";
 
 type GameStatusSectionProps = {
-  gameRows: AgentStatusRow[]
+  gameRows: AgentStatusRow[];
 };
 
 export const GameStatusSection = ({ gameRows }: GameStatusSectionProps) => {
