+++ console/src/AgentStatus/LiveDeliveryStatusSection.tsx
@@ -1,10 +1,10 @@
 import { AgentStatusSectionCard } from "./AgentStatusSectionCard";
-import type { AgentStatusRow } from "./AgentStatusSectionCard";
+import type { AgentStatusRow } from "./types";
 
 export const LIVE_DELIVERY_SECTION_TITLE = "配信状況";
 
 type LiveDeliveryStatusSectionProps = {
-  liveDeliveryRows: AgentStatusRow[]
+  liveDeliveryRows: AgentStatusRow[];
 };
 
 export const LiveDeliveryStatusSection = ({ liveDeliveryRows }: LiveDeliveryStatusSectionProps) => {
