+++ console/src/AgentStatus/createAgentStatusSections.ts
@@ -1,8 +1,8 @@
 import type { AgentStateResponse, AgentStatusSection } from "./types";
 import { createAgentStatusRows } from "./createAgentStatusRows";
-import { GAME_SECTION_TITLE } from "../agentStatusSections/GameStatusSection";
-import { LIVE_DELIVERY_SECTION_TITLE } from "../agentStatusSections/LiveDeliveryStatusSection";
-import { MARKOV_MODEL_SECTION_TITLE } from "../agentStatusSections/MarkovModelStatusSection";
+import { GAME_SECTION_TITLE } from "./GameStatusSection";
+import { LIVE_DELIVERY_SECTION_TITLE } from "./LiveDeliveryStatusSection";
+import { MARKOV_MODEL_SECTION_TITLE } from "./MarkovModelStatusSection";
 
 const LIVE_DELIVERY_ROW_LABELS = ["配信指標", "タイトル", "配信URL", "開始時刻", "発話内容"] as const;
 const MARKOV_MODEL_ROW_LABELS = ["生成N-gram", "これまでの発話"] as const;
