+++ console/src/AgentStatus/index.tsx
@@ -1,15 +1,11 @@
-export { AgentStatus } from "./AgentStatusComponent";
+export { AgentStatus } from "./AgentStatus";
 export { createAgentStatusRows } from "./createAgentStatusRows";
 export { createAgentStatusSections } from "./createAgentStatusSections";
 export {
-  parseAgentStateResponse,
   createMockAgentStateResponse,
   isAgentStateMockQueryEnabled,
   shouldUseMockAgentState,
   startAgentStateAutoRefresh,
-  AGENT_STATE_MOCK_NOTICE_MESSAGE,
   AGENT_STATE_REFRESH_INTERVAL_MS,
-  shouldShowAgentStatusErrorForEventSourceError,
-  INVALID_AGENT_STATE_RESPONSE_ERROR,
+  parseAgentStateResponse,
 } from "./agentStatusState";
-export type { AgentStatusRow, AgentStatusSection, AgentStateResponse } from "./types";
