--- console/src/agentStatusSections/AgentStatusSectionCard.test.tsx
@@ -1,46 +1,0 @@
-import { describe, expect, it } from "bun:test";
-import { renderToStaticMarkup } from "react-dom/server";
-import { AgentStatusSectionCard } from "./AgentStatusSectionCard";
-
-describe("AgentStatusSectionCard", () => {
-  it("renders section title and rows including link rows", () => {
-    const html = renderToStaticMarkup(
-      <AgentStatusSectionCard
-        title="テストセクション"
-        rows={[
-          { label: "状態", value: "配信中" },
-          { label: "配信URL", value: "https://example.com/live", href: "https://example.com/live" },
-          {
-            label: "ゲーム情報",
-            valueComponent: (
-              <ul>
-                <li>status: idle</li>
-              </ul>
-            ),
-          },
-        ]}
-      />,
-    );
-
-    expect(html).toContain("テストセクション");
-    expect(html).toContain("状態");
-    expect(html).toContain("配信中");
-    expect(html).toContain("href=\"https://example.com/live\"");
-    expect(html).toContain("<ul");
-  });
-
-  it("pins row layout to the top instead of stretching across card height", () => {
-    const html = renderToStaticMarkup(
-      <AgentStatusSectionCard
-        title="テストセクション"
-        rows={[
-          { label: "状態", value: "配信中" },
-          { label: "発話内容", value: "こんにちは" },
-        ]}
-      />,
-    );
-
-    expect(html).toContain("content-start");
-    expect(html).toContain("items-start");
-  });
-});
